requirements.txt
mysql-connector-python==8.2.0  # Подключение к MySQL
pandas==2.2.3                  # Работа с данными
python-dotenv==1.1.0           # Работа с переменными окружения
sqlalchemy==2.0.23             # ORM для работы с БД (альтернатива)
requests==2.31.0              # HTTP-запросы для получения данных
phonenumbers-9.0.6          # Работа с номерами телефонов
